requirements.txt
langchain>=0.3.0
langchain-groq>=0.5.0
langchain-tavily>=0.2.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0
chromadb>=0.4.0
sentence-transformers>=3.0.0
pymupdf>=1.23.0
nltk>=3.8.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pydantic>=2.0.0
langchain-core>=0.3.0
